wifi: fix incorrect SoftApCapability when T-on-S
A new overlay configuration:config_wifiDriverSupportedNl80211RegChangedEvent
is introduced from T. It is used to control if SoftApCapability is
needed to be updated when country code changed. The S device should work
normally even if device is using default value.
Thus the fix includes

1. Align the timing to update SoftApCapability between T-on-S and S.
2. New countrycode callback should only work after T.

Also fixed the flaky unit test:
WifiCountryCodeTest#testClientModeManagerAndSoftApManagerDoesntImpactEachOther
in build T-on-R

Bug: 234105280
Bug: 234757198
Test: atest -c FrameworksWifiTests
Change-Id: I6088541e71760194ab13e48f5632af5279083da9