Add flag |isCurrentlyConnected| to WifiConfiguration.
WifiConfigManager#removeExcessNetworks currently checks
whether a config's |status| field is CURRENT to determine
if it connected. However, in the case where we connect to a
network but no internet is available, we can end up in a state
where we are connected to the network, but the config's |status|
is set to DISABLED. This can cause a network's config to get
removed in removeExcessNetworks(), even if it is the currently
connected network.

Instead of using |status| to determine the connection status,
maintain a separate flag |isCurrentlyConnected| which is set
to true on connection and false on disconnection.

Bug: 202358135
Test: atest WifiConfigurationTest  WifiConfigManagerTest
Change-Id: Iba4c70f2233826eb091cfe24c9f404fffc8c9263