wifi: fix incorrect SoftApCapability when T-on-S
A new overlay configuration:config_wifiDriverSupportedNl80211RegChangedEvent
is introduced from T. It is used to control if SoftApCapability is
needed to be updated when country code changed. The S device should
update after set cc succeeded and received cc changed since S build we
don't know if the driver supported or not.

Also bypass the optimized check for CC callback. In S build it should
always to trigger since it is used to update SoftApCapability only.

Also rename mLastActiveCountryCode to mLastReceivedCountryCode since it
is used to handle the set cc is same as mLastReceivedCountryCode case.

Bug: 234105280
Test: atest -c FrameworksWifiTest on T-on-S device.
Test: atest -c FrameworksWifiTest on T build
Test: Manual test wirh command to change country code on S build,
the cc changed event is triggered twice (set and received) as expected.

Change-Id: Iad7888f62a96d9b7a8bd60c21a19a7b0f61f217f