wifi: fix CC didn't reset to default when cellular loss
There are two use cases when calling pickCountryCode.
1. Pick to know the current country code
2. Pick to update country code to driver

Android T supports a callback to indicate what is the country code which
is being used by driver. For no-modem device, the driver country code
might be update by driver (802.11d). So the pick country code should return
the driver country code for this use case (Use case to know the current country code).
But it is not correct for the use case which updates cc to the driver.
When cellular loss, it should fallback to use a default country code
from telephony country code.

Fix: Add a boolean to indicate use case and remove driver country code
from candidate for writing use case.

Bug: 236160004
Test: atest -c FrameworksWifiTests

Change-Id: Idd4c218ab52691499fdc5284d800a38ad6719dfb