eap: allow delayed approval for auto-connected TOFU networks
An unattended TOFU notification blocks Wi-Fi connectivity for a long time,
i.e. when a TOFU notification is not dismissed, this device won't try to
connect to another available network in range.

To prevent this, a delayed approval for auto-connected TOFU network is
introduced. When a TOFU network is auto-connected and a notification is fired,
auto-connect will be disabled and this network will be disconnected after 10
seconds. Once the user accepts it, the auto-connect will be enabled again for
this network.

Bug: 237593725
Bug: 237501731
Test: atest FrameworksWifiTests
Change-Id: Ia1ce34d3e39a32f8d0e85c0f61758421c4910f13