Do not kill WifiDialog process if there are activities remaining
There may be edge cases where WifiDialogActivity finishes and a new
dialog is requested immediately. In this case, the new dialog request
starts a new WifiDialogActivity since the first one is finishing.
However, once the first activity calls onDestroy(), we end up killing
the entire Process which kills the second activity. Instead, we should
check if there are any other activity instances before calling killing
the process.

Bug: 236803960
Test: Enable iface conflict dialog, turn on STA + AP, go to P2P settings
and reject the dialog. Verify the second dialog pops up immediately and
the process is not killed. Verify accepting the second dialog kills the
process.

Change-Id: I19ffebdc1b7285b67dd0b4ca825dd27fac945e2b