Call the getWifiEnabledState when Wifi is enabling.
For system prior to T, make blocking call of the getWifiEnabledState on the
WifiHandlerThread, when the current wifi state is enabling.

Bug:238591443
Test:atest FrameworksWifiTests.
Manual test with T mainline module + T build and T mainline module + S build.
Turn on/off Wifi from the switch of the shortcut key, Wifi should be able to turn on/off successfully.

Change-Id: I741be7ca7abc25740b6acac4c94ff220d3ec394c