Apply user rejection to deferred ICM requests
State machines that are put into the WaitingState by
InterfaceConflictManager waiting for a user reply may still receive
iface request messages. Since these are deferred by the WaitingState
and re-executed once the user replies, we may end up instantly prompting
the user again once the deferred requests are executed. To avoid this,
we should apply a user rejection to all of the deferred requests.

If the user approved, then ICM is expected to immediately return
ICM_EXECUTE_COMMAND since the request should be satisfied by the newly
created iface.

Bug: 239503698
Test: atest InterfaceConflictManagerTest
Change-Id: I37c293af321c26e67471f1e69e6c6710b4a7d251