Reset ICM if wifi turns off before dialog is dismissed
InterfaceConflictManager relies on a reexecuted iface request to stop
waiting for the user response to the iface conflict dialog. If wifi is
turned off or the active clients are removed before the dialog is
replied to, then P2P state machine will ignore the deferred iface
requests after the dialog response is received. In this case, we should
explicitly reset ICM so it doesn't get stuck waiting for the reexecuted
request.

Bug: 239503697
Test: atest WifiP2pServiceImplTest, manually verify
InterfaceConflictManager does not get stuck in waiting state if the P2P
settings is closed before the interface conflict dialog is replied to.

Change-Id: I779d899adc1aa7d4e85ea08042b64965e2139d3f