wifi: Require permission for AP State change intent after SC-V2 only
The security fix ag/16631792 was required to fix after SC-V2.
So add sdk version check to align behavior for mainline user.
i.e. To device which is running with code before SC-V2, use sticky intent.
The device which is running with code after SC-V2, use intent and
require the permission.

Bug: 231214355
Test: atest -c FrameworksWifiTests
Change-Id: Iaaeb5b1d9a875b79c6149589574372b231fa9365