Do not throw exception for setting invalid HTTP proxy
WifiConfiguration.setHttpProxy() throws an IllegalArgumentException
for invalid ProxyInfo. This may crash the main wifi thread in
WifiConfigManager if we copy an invalid ProxyInfo from an
IpConfiguration, which does not check the validity of its ProxyInfo.

Since IpConfiguration does not validate its ProxyInfo, and
WifiConfiguration.setHttpProxy() does not declare its
IllegalArgumentException, we should remove this exception and simply
refuse setting an invalid ProxyInfo.

Bug: 240616051
Test: atest WifiConfigurationTest
Change-Id: Ie6929fac7fc08caa07e3f4357525e8059213da47