Fix uncaught exception when adding config with invalid hex SSID
WifiConfiguration.getProfileKey() uses WifiSsid.fromString(SSID), which
may throw an IllegalArgumentException if the hex SSID is poorly
formatted (i.e. not hex or not even lengthed). This causes a crash in
the main wifi thread when the profile key is used to query any existing
configs during a config add. Instead, we should build the profile key
from the SSID field directly and avoid using WifiSsid.fromString().

Bug: 240616051
Test: atest WifiConfigurationTest WifiConfigManagerTest
Change-Id: Id2d03657712623bd51b7d3aeaca8085901c01597