Use cached hash function for MAC calcuation
Getting the hash function from the keystore is an blocking operation
that could take up to tens of milliseconds. So used the cached version
for as long as possible until the cached handle is invalidated by the
keystore.

Highly optimizes performance when restoring a large number of saved
networks. Also improves performance when connecting to a network that's
using persistent randomized MAC.

Bug: 241166720
Test: atest com.android.server.wifi
Change-Id: I1cf6cf33ae1e651dd5c99a802bc2245d60f04174