Use timeout to clear recorded BSSID charsets
BSSID->Charset mappings are cleared whenever we disconnect from a
network. However, this means the Charset we get for a BSSID may be
incorrect before the next scan results come in. Thus we should use a
timeout mechanism to clear the mappings to ensure that we still have
valid mappings before the next scans.

Bug: 140084570
Test: atest SsidTranslatorTest
Change-Id: I62378bee8ea1794de7e0efced69d63917f7f53ec