Fallback to alternate SSID on network not found notification
One SSID in the framework may be the UTF-8 translation of up
to 2 different raw SSIDs (UTF-8 or an alternate charset).
When we connect to a network, we send one of these raw SSIDs to
supplicant. In case this SSID is not found, fallback to the other raw
SSID to connect to.

Bug: 140084570
Test: atest FrameworksWifiTests, manually connect to a network in the
wifi picker made of a UTF-8 AP and GBK AP, turn off the connected AP,
connectto a different network and then reconnect to the first network,
verify connection succeeds and fallback selection was indicated in the
logs.

Change-Id: I0ff7d6695f3f3fd0b63872857f3fcdea084e82a3