Add required flag to registerReceiver call in WifiDialogManager
Android T added support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting U+ that are registering for
non-system broadcasts. This commit adds the RECEIVER_EXPORTED
flag since the CLOSE_SYSTEM_DIALOGS broadcast is already
limited by the platform but required for some tests.

Bug: 234659204
Test: atest WifiDialogManagerTest
Merged-In: I8fa1039057f859fc27c69b71e6c4227675ccf0a4
Change-Id: I3a843c55a49ef9e59978f5f35ec186dc97e439ee