wifi: Fix improper log level
Reason: The timeout message is posted from alarm manager thread but the configuration (timeout enabled value) or connected client number might be changed at the same time which is too late to cancel timeout message. So it may happen the unexpected configuration when executing timeout message.

So log level should use d or i instead of wtf since it may be happened.

Bug: 242310283
Test: build and TH verified
Change-Id: I7d34530a5de00b606b44027fb34635ea11831e56