Persist autoJoinGlobal across wifi toggle for device admins
If the WifiManager#allowAutojoinGlobal is called directly by a device
admin, or if the original caller is a device admin, then user wifi
toggle should not reset autoJoin back to true.

Also, do not allow non device admins to re-enabled autojoin if it had
been disabled by a device admin.

Bug: 239687722
Test: atest com.android.server.wifi
Change-Id: I65fb8d9b8ffd1f708321bf634ee3f36bdd12ad7d