Replace AsyncChannel with AIDL interface for WifiScanningService.
Currently WifiScanning Service relies on the AsyncChannel to allow
WifiScanner to communicate with the service.

Currently if WifiScanner is obtained in a new context then the new
instance of WifiScanner is created when calling getSystemService.

The AsyncChannel is created in the WifiScanner constructor and
the remote Messenger object is registered and never released.
The WifiScanningService holds strong reference to the Messenger of the
WifiScanner that lead to biner proxy objects leak.

The below changes are introduced:
1. Add AIDL interface to IWifiScanner and the implementation to
   WifiScanning Service.
2. WifiScanner callback is registered to WifiScanningService when
   calling the registerScanListener or startScan.
3. When a single scan is completed, the listener will be removed from
   service.
4. When the client listener's binder object died, the listener will be
   removed from the service.

Bug: 231449688
Bug: 240890134
Test: atest FrameworksWifiTests

Change-Id: I81cc9c3b6ec7d2ea06b22277ada5a7bb96eed475
Merged-In: I81cc9c3b6ec7d2ea06b22277ada5a7bb96eed475