Avoid WifiDialogManager NPE if WifiDialog package name is null
WifiDialog package name is needed to create the intent to launch the
app. The WifiDialog package name is generally available by the time the
framework creates any dialogs, but the method to retrieve the package
name is still @Nullable, which may cause an NPE in unexpected scenarios.
Thus we should avoid the NPE by explicitly handling the null package
name case.

Bug: 243601281
Test: atest WifiDialogManagerTest
Change-Id: Idbf17783ededfdb9f9a04a7d509fb2b4f0ca7131