Fix LOHS app label in ICM dialog
Requestor worksource for LOHS is faked as "Settings" in order to gain
priority over STA (legacy LOHS behavior). However, this results in the
ICM dialog showing the app label of "Settings" instead of the actual app
that requested LOHS. Since ICM dialog enabled bypasses the worksource
priority scheme already, we should NOT fake the LOHS worksource if the
dialogs are enabled.

Bug: 239364438
Test: enable ICM dialog, turn on P2P, use test app to request LOHS and
verify ICM dialog shows the test app name instead of "Settings".

Change-Id: I6f61433dc9421ab24809a965d32dc2e5492c954e