Do not auto-open keyboard on P2P pin in landscape
The P2P invitation dialog auto-opens the keyboard on the pin input, but
landscape mode on phones is not big enough to display both the P2P
dialog and the keyboard at same time. Thus we shouldn't auto-open the
keyboard in landscape mode for less-than-large sized screens (tablets),
and instead have the user manually tap on the pin field to use the
full-screen landscape mode keyboard.

Bug: 248187548
Test: adb shell cmd wifi launch-dialog-p2p-invitation-received
DeviceName -p, manually verify keyboard does not auto-open in landscape
mode

Change-Id: I5f20db226061e34205a146e4af238356e340d334