Fix WifiConfiguration.hasNoInternetAccess()
WifiConfiguration.numInternetAccessReports is incorrectly reset to 0
on a failed validation, which causes hasNoInternetAccess() to return
true. We should only reset numInternetAccessReports to 0 on a successful
validation.

Also, it seems numInternetAccessReports=0 returns
hasNoInternetAccess()=true for never-connected networks, but since this
value isn't written to the config store, we should use
hasEverConnected() instead.

Bug: 243477918
Test: atest FrameworksWifiTests

Change-Id: I287cbe82af168af04d3f4ff8d01543654e8bcfb3