Update Layer2 Information before start provisioning
L2Key information needs to be updated to ipclient before
starting the provisioning. This is needed for network stack
to check the DHCP lease cache and trigger DHCP request instead
of DHCP discover, to the network which was connected before.

Bug: 251367593
Test: atest com.android.server.wifi.ClientModeImplTest#testLayer2InformationUpdate
Test: Tested connect-disconnect to same AP multiple times and confirmed
from the sniffer logs that only initial connection trigger DHCP discover
packet and subsequent connections are initiating DHCP by sending  DHCP request
packets.

Change-Id: Ib370c4c9f02f9b3b948023ac7ab4cb47d5b40dc7