Avoid overwriting the scan result of a BSS with Hidden SSID scan result
The results of the recent scan results are stored in the framework as a map of
bssid -> ScanResult. If a hidden network is configured, wificond may
report two scan results for the same BSS, ie. One with the SSID and
another one without SSID. If the second scan result is hidden SSID scan
result, then the first one with SSID gets overwritten & application gets
an empty SSID. So avoid overwriting the scan result from same BSSID by
checking if the scan result with SSID is already written to Scan result
hash map.

Bug: 256587747
Test: atest com.android.server.wifi
Test: # adb shell cmd wifi list-scan-results
      and confirmed from the results that hidden SSID scan result
      is not overwriting the scan result with SSID.
Change-Id: Ie7ae925e0df34fc10b64e28f7533001022158133