Set mWifi to null after HalDeviceManager calls stop().
Immediately following a call to HalDeviceManager.stop(),
WifiVendorHal checks HalDeviceManager.isReady().

In the old version of HDM, this call to isReady()
happened before mWifi was set to null, so it would return
true. However, in the restructured version of HDM, we
set mWifi to null first, meaning that isReady() returns
false.

This triggers an error handler which attempts to
disable wifi. In the CTS tests, the error handler
will run at the same time as the test is trying
to enable wifi, causing the wifi to get disabled
right as it is trying to start up.

Bug: 257044385
Bug: 257263488
Test: atest WifiManagerTest#testTetheredBridgedAp
Change-Id: Ice86f1b16954749880d6d58fe3d8f13345bc7ee4