Clean up WifiDialogActivity's home button logic
WifiDialogActivity cancels all dialogs when ACTION_CLOSE_SYSTEM_DIALOGS
is received. However, this happens before the device transitions to the
home screen, allowing apps to immediately request a new dialog right
before they're put into the background. This results in a dialog popping
over the home screen.

Instead, we should cancel the dialogs in onStop() (if not a
configuration change or screen off), which happens after the user navigates
to the home screen and the rest of the activities are put to the
background.

Bug: 257398441
Test: enable ICM dialogs and use test app to constantly request an iface
dialog, verify home button press closes the dialog and does not pop a
dialog over the home screen.

Change-Id: I25ff807307d2d1c4d609df9d316f91e3f1407dbc