Make dialog handles @NonNull
WifiDialogManager may return null dialog handles if some of the
arguments are illegal. However, if these null handles are uncaught then
this may cause an NPE on the wifi thread. Thus it is better to simply
log the illegal argument and display the dialog without the illegal
arguments anyway.

Bug: 260544149
Test: atest WifiDialogManagerTest, manually verify dialogs via adb shell
cmd wifi launch-dialog* commands

Change-Id: I7ae43d4b758e35748311b093e904432c2d3a429c