[Aware] Rename flag to clusterEventType and mac to clusterId
In WifiAwareClientState and WifiAwareStateManager, the term flag is
used to indicate the cluster event type that triggered the callback.
To be more descriptive, the naming of clusterEventType could be used.

Secondly, the cluster Id after propogated from state manager to client
state, it became mac. Though the cluster id is indeed a mac address, to
avoid confusion, it could be renamed to its origin - cluster ID.

Bug: 263948533
Bug: 263949214
Test: atest com.android.server.wifi
Change-Id: Ia255edf68b3bc8fcb7e9712d178c9b4ea527a195