Use scan detailed cache to get MLO info
ScanRequestProxy#getScanResult(bssid) returns scan results cached from
the last full single scan request. If the MLO link AP is missed on last
scan or last scan is not a full scan, ScanRequestProxy will not have
the scan results corresponding to the link AP. This results in MLO
association with no affiliated link info.

WifiConfigManager has the detailed scan cache which is updated
for every full and partial scans. Use that instead of ScanRequestProxy.

Bug: 264228374
Test: atest com.android.server.wifi
Change-Id: I85c41744967b8cdac3e0f70b8ae121e32232d272