[TOFU] Handle anonymous identity correctly
If anonymous identity is not provided by the user, TOFU connection
would fail because wpa_supplicant doesn't have any identity to
start the EAP authentication with, since it was deleted as part of
the new TOFU flow. If anonymous identity is not provided by the
user, use "anonymous" rather than the real user identity, to
prevent untrusted servers from tracking real user identities.
Note: As per RFC3748, servers should include a method-specific
mechanism for obtaining the real identity, so that they do not
have to rely on this anonymous identity, in other words, it
should be ignored.

Bug: 251910611
Bug: 267633332
Test: Integration test with WPA-Enterprise setup
Test: atest InsecureEapNetworkHandlerTest
Change-Id: I57d7cc09261186fd470acef224a74da10d129da5