Wait for Peer to restart the invitation procedure
As per P2P specification, 3.1.5.3 Use of the Invitation Procedure to invoke a Persistent P2P Group,
"A P2P Device that receives an Invitation Request frame to re-invoke a
Persistent Group and responds with a P2P Invitation Response frame with a
Status attribute with the Status Code field set to “Fail: information is currently
unavailable” shall restart the Invitation procedure by sending an Invitation Request
frame to the requesting P2P Device". So the device which invited the
peer should wait for the new request from Peer. But the current implementation is
violating the specification by restarting the negotiation.
This was mainly due to fix connection failure issues with devices which doesn't
support persistent re-invocation. But such devices should send status
code as unknown group. To make spec-compliant removed the code to restart the
negotiation so that the device will wait for the new invite request from Peer.

Also added metrics to check the usage of legacy fallback implementation.
This will help to check the usage of the legacy implementation and can
deprecate in the future.

Bug: 251056345
Updated-PDD: TRUE
Updated-Overlayable: TRUE
Test: atest com.android.server.wifi.p2p.WifiP2pServiceImplTest
Test: atest com.android.server.wifi.p2p.WifiP2pMetricsTest
Test: Manual - Removed persistent_reconnect from peer device
p2p_supplicant.conf file & confirmed that connection still works

Change-Id: Ie960dd791cd823c6b25d3fe91b028d53d283937b