disable unwanted network temporarily
Add optional feature for devices using the AOSP scorer.
When the wifi is not good because of rssi or others, and the
score is less than 50, the default network will change to data,
however, the wifi will always try to auto connect if the rssi
beacomes strong, this will make the network is not stable. This feature
will cause the unwanted network to be temporarily disabled to provide
more stability to the cell connection.

Devices using the external connected scorer likely won't need this
feature since the external scorer already can trigger network disable.

Updated-Overlayable: TRUE
Test: atest com.android.server.wifi
Test: Manually verified
Bug: 269412249
Change-Id: I85c0bf3d5e1b48818f7e84fb2387bd4b59c6f595
Signed-off-by: zhujiatai <zhujiatai@xiaomi.com>