Handle Auth timeout due to wrong password
Some AP implementations doesn't send de-authentication or dis-association
frame after EAPOL failure. This leads to authentication timeout in supplicant.
If this network was not connected before, the authentication timeout may be due
to wrong password. So on receiving authentication timeout,
if the failure is seen in very first connect attempt,
mark it as a authentication failure due to wrong password.

Bug: 263000302
Test: atest com.android.server.wifi
Test: Manual - Wrong password notification is seen in UI when supplicant
reports authentication failure in first connect attempt.

Change-Id: I6baf078f5a2d3956d4092a355711c7ba8bc285eb