Forcefully trigger the join existing group method
Added wifiP2pConfig builder to specify that the device
wants to join in an existing group as client.
Usually group owner sets the group owner capability bit in beacons/probe responses. But
there are devices out in the field which doesn't set the group owner capability bit.
The new API is for applications which can get the peer group owner capability via OOB
(out of band) mechanisms and forcefully trigger the join existing group logic.

Bug: 208312959
Test: atest WifiP2pConfigTest
Test: atest WifiP2pServiceImplTest
Change-Id: I589b3170bf915a806c53c53d1f1c70a2f58f6707