Fix MLO roaming
If the FW roams to a new AP, the BSSID may not be in the scan detailed
cache. Set Multi-Link info only if the BSSID is present in the scan
detailed cache and AP is a Multi-Link Device. If not, clear Multi-
Linkinfo from WifiInfo.

After association ClientModeImpl#updateMloLinkAddrAndStates() will
update the Multi-Link info from supplicant.

Bug: 270089496
Test: atest com.android.server.wifi.ClientModeImplTest
Change-Id: I96442171bd360b9aea0a8f9b7b2844cfc5ac5c0d