Refactor handleIpReachabilityFailure in ClientModeImpl.
To do the similar logic for reachability failures which are triggered
from WiFi RSSI polling or organic kernel probe to the failure post
roaming and try to remove the nested if-else condition check(e.g. we
may need to read the overlay resource to decide if the new behavior
of handling other failures is enabled), move the common code to a
separate function for different loss reasons, this refacotr makes
logic clear and easier to add new condition check later. This CL should
be a no-op change, just refactoring the code.

Bug: 233684075
Test: atest com.android.server.wifi
Change-Id: I8dddf8dd4df881aa3f6df49d61ac65c7c861401a