[TOFU] Fix flow for EAP-TLS
Fix the TOFU flow for EAP-TLS, when using WPA2/3-Enterprise and
WPA3-Enterprise in 192-bit mode. The latter configures RSA 3K or
ECDSA security modes based only on the user cert in the TOFU flow
because the CA cert is not available. Do not remove the user
certificate during TOFU flow, which is required for the TLS
handshake.

Bug: 269218756
Test: atest WifiKeyStoreTest InsecureEapNetworkHandlerTest
Test: Use TOFU in a network using EAP-TLS
Change-Id: Ife5de732f3d5a3c235433860007d17d66bfd942f