Make restoreSoftApBackupData non-blocking
WifiManager.set/getSoftApConfiguration() was made non-blocking, but
restoreSoftApBackupData still posts to the wifi thread. This means apps
may still get the old config if they call restoreSoftApBackupData and
immediately call getSoftApConfiguration() before the actual update is
run on the wifi thread. To fix this, make restoreSoftApBackupData run
entirely on the binder thread too.

Bug: 269899499
Test: atest WifiServiceImplTest android.net.wifi.cts.WifiBackupRestoreTest#testCanRestoreSoftApBackupData
Change-Id: I7d1fa4825c3a673bef6369547ddeb0a964299ba1