Filter wtf logging for External Scorer
During MBB there is a very brief period (at the time when the old
primary transitioned to secondary, but new primary not transitioned yet)
where the external scorer is not linked with any ClientModeManager, and
mCallback is null.

Also, there are possibly cases where the external scorer is
unregistered, but scorer still triggered the callback.

Limit wtf logging to only happen after 3 consecutive times to filter out
the above cases which are not real failures.

Bug: 242310511
Test: atest com.android.server.wifi

Change-Id: Ieb3c6f849c687feca2aa6c3031e60393d696c31c