Keep WifiDialogs up even when screen is off
Do not cancel dialogs in onStop() anymore, which allows dialogs to show
even when the screen is off and the Activity is stopped. Cancelling in
onStop() was needed because other apps might cover up active dialogs
without the user knowing, but using TYPE_SYSTEM_DIALOG now shows dialogs
over all apps and this behavior isn't needed anymore.

Bug: 239474988
Test: adb shell cmd wifi launch-dialog-simple -t Title -m Message,
manually verify dialogs stay up when screen is off, and remain on top if
Settings is opened while a dialog is up.

Change-Id: I10dbcdbfc31990e36a63e4c994e39a723ed0aaa6