Fix a frequent print on getWifiChipCapabilities()
The mWifiNative.getMaxMloStrLinkCount() method is called for every
network selection to check if the device is capable of STR MLO. The
information is cached at WifiVendorHal, but the cache needs at least one
call to succeed from the chip.

For a chip that does not implement this method, the "WifiChipAidlImpl:
getWifiChipCapabilities failed with service-specific exception:
android.os.ServiceSpecificException:  (code 4)" error will be frequently
printed.

To fix this, the chip method should only be invoked once during
interface creation.

Bug: 274770910
Test: atest com.android.server.wifi.WifiVendorHalTest#testGetChipCapabilities
Change-Id: I4ba4ff302ddf901b3f8af7daef1175156d99a147