Flush the PMK cache if credential is changed
This CL includes two changes,
1. Revert of commit - Ie00dc84147780c634e2a69a2e9aae22e12ccaa6a
as pseudonym changes for every connection and the change in
a pseudonym cannot be considered as a credential change. The original
code was added to reload the network configuration in supplicant
with the decorated pseudonym. But recent implementation already
updates the decorated psuedonym in supplicant as soon as the connection
complete event comes from the supplicant.
2. Flush the PMK cache only when the credential is changed.
This is a slight modification of fix added in
commit - I1369978c9af1bb75ff76edd3b7ba8b22f921884c

Bug: 256786992
Test: atest com.android.server.wifi
Test: Manual - connect/disconnect SIM based authentication network
      and confirmed that reconnection doesn't go through full EAP
      authentication.
Change-Id: If763e92c5fb8b111197a3d6edc26cdff2416b24e