Remove the band check if primary STA is MLO
Android 13 or higher devices can connect to two unrestricted Wi-Fi
networks (APs) at the same time and use both for internet access. If
WifiManager#WIFI_MULTI_INTERNET_MODE_MULTI_AP is enabled, secondary
STA can connect to an AP which is not in the same band as primary.

If the primary STA is a Multi-Link Operation (MLO), it can have links
in multiple bands. Therefore, for the secondary STA, pick any
candidates other than the BSSIDs that are affiliated with the primary
STA. The firmware will adjust the multi-links accordingly

Bug: 277606046
Test: atest com.android.server.wifi.WifiConnectivityManagerTest
Change-Id: I3ae1ddd6cef525004983527055d8c2c94140e72a