Retain auto-upgraded flag when updating a config.
The auto-upgrade flag for a config gets automatically set to false when
the user adds/updates an upgradable config. However, since the user may
be updating a config that was originally added by auto-upgrade, this
behavior wrongly changes the flag from true to false.

Instead of always setting the flag to false, we should retain whatever
value was set by the external config.

Bug: 262312532
Test: atest WifiConfigManagerTest, manually verify adding a PSK network
and then modifying the network in Network details won't set the SAE
auto-upgrade flag to false

Change-Id: I2f1705fccaa45fdba44af0a4c8578496b391a30a