Fix cached usable channels for multi-band bitmap
Multi-band bitmaps like WIFI_BAND_24_5_WITH_DFS are allowed in
WifiManager.getUsableChannels(), but the current implementation for
getting cached channels checks for equality instead of the individual
bit value. Fix this by making sure we check each band's bit
individually and returning the channels for all of the bands specified.

Bug: 246244842
Test: atest WifiServiceImplTest
Change-Id: I78ea3a116062353e48784d5c407bed02f14ff8e5