Skip binder service check for supplicant died event
The commands to supplicant can fail with remote
exception when supplicant is in the middle of termination
(Supplicant restart/crash). This is due to race condition between
two threads, framework runs in the wifi main thread
and supplicant service runs in the binder thread/eloop thread context.
The remote exception clears the ISupplicant instance. The supplicant
binder death handler should not be blocked for this reason, as the handler
doesn't depend on ISupplicant instance. The event should be handled to restart wifi.

Bug: 277171008
Test: Manual - Connect STA to WPA3 AP & kill supplicant via
      adb shell killall wpa_supplicant command. Confirmed that
      the framework is able to restart supplicant and connect back
      to same AP.
Change-Id: I08d8d05b9e8f5c8264b8930ff09710b5cadba6ac